Add docs for OpenSearch
