+++ src/FilterService/FilterType/SearchIndex/NumberRange.php
@@ -14,17 +14,14 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\OpenSearch;
+namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\SearchIndex;
 
-use Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\AbstractFilterType;
 use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\ProductList\ProductListInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractFilterDefinitionType;
-use Pimcore\Model\DataObject\Fieldcollection\Data\FilterNumberRangeSelection;
+use Pimcore\Model\DataObject\Fieldcollection\Data\FilterNumberRange;
 
-/**
- * @deprecated This class will be moved to the SearchIndex namespace in version 2.0.0.
- */
-class NumberRangeSelection extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\NumberRangeSelection
+
+class NumberRange extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\NumberRange
 {
     public function prepareGroupByValues(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList): void
     {
@@ -32,33 +29,27 @@ class NumberRangeSelection extends \Pimcore\Bundle\EcommerceFrameworkBundle\Filt
     }
 
     /**
-     * @param FilterNumberRangeSelection $filterDefinition
+     * @param FilterNumberRange $filterDefinition
      *
      */
     public function addCondition(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList, array $currentFilter, array $params, bool $isPrecondition = false): array
     {
         $field = $this->getField($filterDefinition);
-        $rawValue = $params[$field] ?? null;
+        $value = $params[$field] ?? null;
 
-        if (!empty($rawValue) && $rawValue != AbstractFilterType::EMPTY_STRING) {
-            $values = explode('-', $rawValue);
-            $value['from'] = trim($values[0]);
-            $value['to'] = trim($values[1]);
-        } elseif ($rawValue == AbstractFilterType::EMPTY_STRING) {
-            $value = null;
-        } else {
+        if (empty($value)) {
             $value['from'] = $filterDefinition->getPreSelectFrom();
             $value['to'] = $filterDefinition->getPreSelectTo();
         }
 
         $currentFilter[$field] = $value;
 
-        if (!empty($value) && ($value['from'] !== null || $value['to'] !== null)) {
+        if (($value['from'] !== null || $value['to'] !== null) && ($value['from'] !== '' || $value['to'] !== '')) {
             $range = [];
-            if (!empty($value['from'])) {
+            if (strlen((string)$value['from']) > 0) {
                 $range['gte'] = $value['from'];
             }
-            if (!empty($value['to'])) {
+            if (strlen($value['to']) > 0) {
                 $range['lte'] = $value['to'];
             }
             $productList->addCondition(['range' => ['attributes.' . $field => $range]], $field);
