+++ src/FilterService/FilterType/SearchIndex/NumberRangeSelection.php
@@ -14,16 +14,15 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\OpenSearch;
+namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\SearchIndex;
 
+use Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\AbstractFilterType;
 use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\ProductList\ProductListInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractFilterDefinitionType;
-use Pimcore\Model\DataObject\Fieldcollection\Data\FilterNumberRange;
+use Pimcore\Model\DataObject\Fieldcollection\Data\FilterNumberRangeSelection;
 
-/**
- * @deprecated This class will be moved to the SearchIndex namespace in version 2.0.0.
- */
-class NumberRange extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\NumberRange
+
+class NumberRangeSelection extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\NumberRangeSelection
 {
     public function prepareGroupByValues(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList): void
     {
@@ -31,27 +30,33 @@ class NumberRange extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService
     }
 
     /**
-     * @param FilterNumberRange $filterDefinition
+     * @param FilterNumberRangeSelection $filterDefinition
      *
      */
     public function addCondition(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList, array $currentFilter, array $params, bool $isPrecondition = false): array
     {
         $field = $this->getField($filterDefinition);
-        $value = $params[$field] ?? null;
+        $rawValue = $params[$field] ?? null;
 
-        if (empty($value)) {
+        if (!empty($rawValue) && $rawValue != AbstractFilterType::EMPTY_STRING) {
+            $values = explode('-', $rawValue);
+            $value['from'] = trim($values[0]);
+            $value['to'] = trim($values[1]);
+        } elseif ($rawValue == AbstractFilterType::EMPTY_STRING) {
+            $value = null;
+        } else {
             $value['from'] = $filterDefinition->getPreSelectFrom();
             $value['to'] = $filterDefinition->getPreSelectTo();
         }
 
         $currentFilter[$field] = $value;
 
-        if (($value['from'] !== null || $value['to'] !== null) && ($value['from'] !== '' || $value['to'] !== '')) {
+        if (!empty($value) && ($value['from'] !== null || $value['to'] !== null)) {
             $range = [];
-            if (strlen((string)$value['from']) > 0) {
+            if (!empty($value['from'])) {
                 $range['gte'] = $value['from'];
             }
-            if (strlen($value['to']) > 0) {
+            if (!empty($value['to'])) {
                 $range['lte'] = $value['to'];
             }
             $productList->addCondition(['range' => ['attributes.' . $field => $range]], $field);
