+++ src/FilterService/FilterType/SearchIndex/SelectFromMultiSelect.php
@@ -14,27 +14,24 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\OpenSearch;
+namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\SearchIndex;
 
-use Pimcore\Bundle\EcommerceFrameworkBundle\Exception\InvalidConfigException;
 use Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\AbstractFilterType;
 use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\ProductList\ProductListInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractFilterDefinitionType;
-use Pimcore\Model\DataObject\Fieldcollection\Data\FilterInputfield;
 
-/**
- * @deprecated This class will be moved to the SearchIndex namespace in version 2.0.0
- */
-class Input extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\Input
+
+class SelectFromMultiSelect extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\SelectFromMultiSelect
 {
+    public function prepareGroupByValues(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList): void
+    {
+        $productList->prepareGroupByValues($this->getField($filterDefinition), true);
+    }
+
     public function addCondition(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList, array $currentFilter, array $params, bool $isPrecondition = false): array
     {
         $field = $this->getField($filterDefinition);
-
-        if (!$filterDefinition instanceof FilterInputfield) {
-            throw new InvalidConfigException('invalid config');
-        }
-        $preSelect = $filterDefinition->getPreSelect();
+        $preSelect = $this->getPreSelect($filterDefinition);
 
         $value = $params[$field] ?? null;
         $isReload = $params['is_reload'] ?? null;
@@ -45,15 +42,14 @@ class Input extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\Filte
             $value = $preSelect;
         }
 
-        if (is_string($value)) {
+        if (!empty($value)) {
             $value = trim($value);
         }
 
         $currentFilter[$field] = $value;
 
         if (!empty($value)) {
-            $value = '.*"' . $value .  '".*';
-            $productList->addCondition(['regexp' => ['attributes.' . $field => $value]], $field);
+            $productList->addCondition(['term' => ['attributes.' . $field => $value]], $field);
         }
 
         return $currentFilter;
